+++ Optimizer/AboutForm.cs
@@ -87,10 +87,7 @@ namespace Optimizer
             string s7 = "deadmoo";
             string s8 = "deadmoon";
             string s9 = "deadmoon © ";
-            string s10 = "deadmoon © 2";
-            string s11 = "deadmoon © 20";
-            string s12 = "deadmoon © 202";
-            string s13 = "deadmoon © 2021";
+            string s10 = "deadmoon © ∞";
 
             switch (l2.Text)
             {
@@ -123,19 +120,9 @@ namespace Optimizer
                     break;
                 case "deadmoon © ":
                     l2.Text = s10;
-                    break;
-                case "deadmoon © 2":
-                    l2.Text = s11;
-                    break;
-                case "deadmoon © 20":
-                    l2.Text = s12;
-                    break;
-                case "deadmoon © 202":
-                    l2.Text = s13;
                     t2.Stop();
-                    //t1.Start();
                     break;
-                case "deadmoon © 2021":
+                case "deadmoon © ∞":
                     l2.Text = s0;
                     break;
             }