+++ Optimizer/Program.cs
@@ -13,7 +13,7 @@ namespace Optimizer
         // Enter current version here
 
         internal readonly static float Major = 6;
-        internal readonly static float Minor = 0;
+        internal readonly static float Minor = 1;
 
         internal static string GetCurrentVersionTostring()
         {